+++ src/httpd/handlers.rs
@@ -71,6 +71,9 @@ mod tests {
         let state = AppState {
             exporter:   exporter.clone(),
             index_page: "Test Body".into(),
+
+            #[cfg(feature = "auth")]
+            basic_auth_config: Default::default(),
         };
 
         let data = Data::new(state);