+++ src/main.rs
@@ -21,6 +21,19 @@ use errors::ExporterError;
 use exporter::Exporter;
 use file::FileExporter;
 
+#[cfg(feature = "bcrypt_cmd")]
+use dialoguer::Password;
+
+#[cfg(feature = "auth")]
+use httpd::auth::BasicAuthConfig;
+
+#[cfg(feature = "bcrypt_cmd")]
+use rand::{
+    distributions::Alphanumeric,
+    thread_rng,
+    Rng,
+};
+
 #[macro_use]
 mod macros;
 
@@ -66,6 +79,60 @@ fn is_running_as_root<U: Users>(users: &mut U) -> Result<(), ExporterError> {
     }
 }
 
+#[cfg(feature = "bcrypt_cmd")]
+// Handles hashing and outputting bcrypted passwords for the bcrypt sub
+// command.
+fn bcrypt_cmd(matches: &clap::ArgMatches) -> Result<(), ExporterError> {
+    // Cost argument is validated and has a default, we can unwrap right
+    // away.
+    let cost: u32 = matches.value_of("COST")
+        .expect("no bcrypt cost given")
+        .parse()
+        .expect("couldn't parse cost as u32");
+    let random = matches.is_present("RANDOM");
+
+    // If a password was given on the CLI, just unwrap it. If none was given,
+    // we either generate a random password or interactively prompt for it.
+    let password = match matches.value_of("PASSWORD") {
+        Some(password) => password.into(),
+        None           => {
+            if random {
+                // length was validated by the CLI, we should be safe to
+                // unwrap and parse to usize here.
+                let length: usize = matches.value_of("LENGTH")
+                    .expect("no password length given")
+                    .parse()
+                    .expect("couldn't parse length as usize");
+
+                thread_rng()
+                    .sample_iter(&Alphanumeric)
+                    .take(length)
+                    .map(char::from)
+                    .collect()
+            }
+            else {
+                Password::new()
+                    .with_prompt("Password")
+                    .with_confirmation(
+                        "Confirm password",
+                        "Password mismatch",
+                    )
+                    .interact()?
+            }
+        },
+    };
+
+    let hash = bcrypt::hash(&password, cost)?;
+
+    if random {
+        println!("Password: {}", password);
+    }
+
+    println!("Hash: {}", hash);
+
+    Ok(())
+}
+
 #[cfg(feature = "rc_script")]
 fn output_rc_script() {
     debug!("Dumping rc(8) script to stdout");
@@ -89,6 +156,14 @@ async fn main() -> Result<(), ExporterError> {
     // If we have been asked to dump the rc(8) script, do that, and exit.
     if matches.is_present("RC_SCRIPT") {
         output_rc_script();
+
+        ::std::process::exit(0);
+    }
+
+    #[cfg(feature = "bcrypt_cmd")]
+    // If we have the auth feature, we can bcrypt passwords for the user.
+    if let Some(subcmd) = matches.subcommand_matches("bcrypt") {
+        bcrypt_cmd(&subcmd)?;
 
         ::std::process::exit(0);
     }
@@ -130,13 +205,26 @@ async fn main() -> Result<(), ExporterError> {
     })?.to_owned();
     debug!("web.telemetry-path: {}", telemetry_path);
 
-    let exporter = Box::new(Exporter::new());
-
-    // Configure and run the http server.
-    httpd::Server::new()
+    // Start configuring HTTP server.
+    // unused_mut here silences a warning if the crate is compiled without auth
+    // support.
+    #[allow(unused_mut)]
+    let mut server = httpd::Server::new()
         .bind_address(bind_address)
-        .telemetry_path(telemetry_path)
-        .run(exporter).await?;
+        .telemetry_path(telemetry_path);
+
+    #[cfg(feature = "auth")]
+    // Set the configuration file for HTTP Basic Auth
+    {
+        if let Some(path) = matches.value_of("WEB_AUTH_CONFIG") {
+            let config = BasicAuthConfig::from_yaml(&path)?;
+
+            server = server.auth_config(config);
+        }
+    }
+
+    let exporter = Box::new(Exporter::new());
+    server.run(exporter).await?;
 
     Ok(())
 }