+++ CHANGELOG.md
@@ -3,13 +3,20 @@
 ## v0.14.0
 
   - Update dependencies.
-  - The MSRV has been bumped to 1.42.0 as required by [actix-web].
+  - The MSRV has been bumped to 1.44.0 as required by dependencies.
   - Fix some minor [clippy] issues.
   - exporter: Remove a clone from metric bookkeeping.
   - exporter: Avoid using clone when creating exporter metrics struct.
   - Change environment variables, removing the `JAIL_EXPORTER_` prefix.
-  - Added `rc_script` feature enabling a `--rc-script` CLI flag which outputs
+  - Added `rc_script` feature, enabling a `--rc-script` CLI flag which outputs
     the `jail_exporter` [`rc(8)`] script on stdout.
+  - Added `auth` feature, enabling HTTP Basic authentication.
+    - Configuration for the HTTP Basic authentication is via a YAML
+      configuration file, the location of the configuration is specified via
+      the `--web.auth-config` CLI argument.
+  - Added `bcrypt` subcommand when compiled with the `auth` feature.
+    - This is to assist users when they're enabling HTTP Basic Authentication
+      as they may not have tools installed for generating bcrypt hashes.
 
 ## v0.13.0